+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2021-04-06
+### Changed
+- Updated project to work with Immersal SDK 1.12.0
+- Should now also work with Unity 2020.2 and Magic Leap XR Plugin 6+
+
+### Added
+- GeoPose localize function
+
 ## [1.11.2] - 2021-01-15
 ### Added
 - Native plugin for Lumin